+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - The third page now redirects to the install page when clicking "Next".
 - A Finish page has been created.
+- If the Installer ran into an error, it would redirect to a fail page.
 
 ## Version 0.0.4.8_proto